+++ src/database.rs
@@ -404,7 +404,6 @@ pub async fn update_games_by_system_id_mark_complete(
             FROM roms r
             WHERE r.game_id = games.id
             AND r.romfile_id IS NULL
-            AND r.parent_id IS NULL
         )
         ",
         system_id,
@@ -460,7 +459,6 @@ pub async fn update_games_by_system_id_mark_incomplete(
             FROM roms r
             WHERE r.game_id = games.id
             AND r.romfile_id IS NULL
-            AND r.parent_id IS NULL
         )
         ",
         system_id,
@@ -511,7 +509,6 @@ pub async fn update_games_mark_incomplete(connection: &mut SqliteConnection) {
             SELECT r.id
             FROM roms r
             WHERE r.game_id = games.id
-            AND r.parent_id IS NULL
             AND r.romfile_id IS NULL
         )
         ",
@@ -1214,11 +1211,11 @@ pub async fn find_roms_without_romfile_by_size_and_md5_and_system_id(
     })
 }
 
-pub async fn find_roms_without_romfile_by_size_and_md5_and_game_name_and_system_id(
+pub async fn find_roms_without_romfile_by_size_and_md5_and_game_names_and_system_id(
     connection: &mut SqliteConnection,
     size: u64,
     md5: &str,
-    game_names: Vec<&str>,
+    game_names: &Vec<&str>,
     system_id: i64,
 ) -> Vec<Rom> {
     let size = i64::try_from(size).unwrap();
@@ -1247,6 +1244,46 @@ pub async fn find_roms_without_romfile_by_size_and_md5_and_game_name_and_system_
             panic!(
             "Error while finding roms with size {} and MD5 {} and game names {:?} and system id {}",
             size, md5, game_names, system_id
+        )
+        })
+}
+
+pub async fn find_roms_without_romfile_by_name_and_size_and_md5_and_game_names_and_system_id(
+    connection: &mut SqliteConnection,
+    name: &str,
+    size: u64,
+    md5: &str,
+    game_names: &Vec<&str>,
+    system_id: i64,
+) -> Vec<Rom> {
+    let size = i64::try_from(size).unwrap();
+    let md5 = md5.to_lowercase();
+    let sql = format!(
+        "
+        SELECT r.id, r.name, r.bios, r.size, r.crc, r.md5, r.sha1, r.rom_status, r.game_id, r.romfile_id, r.parent_id
+        FROM roms AS r
+        JOIN games AS g ON r.game_id = g.id
+        WHERE r.romfile_id IS NULL
+        AND r.name = '{}'
+        AND r.size = {}
+        AND r.sha1 = '{}'
+        AND g.name IN ({})
+        AND g.system_id = {}
+        ORDER BY r.name
+        ",
+        name,
+        size,
+        md5,
+        game_names.iter().map(|game_name| format!("'{}'", game_name)).join(","),
+        system_id,
+    );
+    sqlx::query_as::<_, Rom>(&sql)
+        .fetch_all(connection)
+        .await
+        .unwrap_or_else(|_| {
+            panic!(
+            "Error while finding roms with name {} and size {} and MD5 {} and game names {:?} and system id {}",
+            name, size, md5, game_names, system_id
         )
         })
 }
@@ -1322,7 +1359,7 @@ pub async fn find_roms_without_romfile_by_size_and_sha1_and_game_names_and_syste
     connection: &mut SqliteConnection,
     size: u64,
     sha1: &str,
-    game_names: Vec<&str>,
+    game_names: &Vec<&str>,
     system_id: i64,
 ) -> Vec<Rom> {
     let size = i64::try_from(size).unwrap();
@@ -1353,6 +1390,46 @@ pub async fn find_roms_without_romfile_by_size_and_sha1_and_game_names_and_syste
             size, sha1, game_names, system_id
         )
     })
+}
+
+pub async fn find_roms_without_romfile_by_name_and_size_and_sha1_and_game_names_and_system_id(
+    connection: &mut SqliteConnection,
+    name: &str,
+    size: u64,
+    sha1: &str,
+    game_names: &Vec<&str>,
+    system_id: i64,
+) -> Vec<Rom> {
+    let size = i64::try_from(size).unwrap();
+    let sha1 = sha1.to_lowercase();
+    let sql = format!(
+        "
+        SELECT r.id, r.name, r.bios, r.size, r.crc, r.md5, r.sha1, r.rom_status, r.game_id, r.romfile_id, r.parent_id
+        FROM roms AS r
+        JOIN games AS g ON r.game_id = g.id
+        WHERE r.romfile_id IS NULL
+        AND r.name = '{}'
+        AND r.size = {}
+        AND r.sha1 = '{}'
+        AND g.name IN ({})
+        AND g.system_id = {}
+        ORDER BY r.name
+        ",
+        name,
+        size,
+        sha1,
+        game_names.iter().map(|game_name| format!("'{}'", game_name)).join(","),
+        system_id,
+    );
+    sqlx::query_as::<_, Rom>(&sql)
+        .fetch_all(connection)
+        .await
+        .unwrap_or_else(|_| {
+            panic!(
+            "Error while finding roms with name {} and size {} and SHA1 {} and game names {:?} and system id {}",
+            name, size, sha1, game_names, system_id
+        )
+        })
 }
 
 pub async fn count_roms_with_romfile_by_size_and_sha1_and_system_id(
@@ -1427,7 +1504,7 @@ pub async fn find_roms_without_romfile_by_size_and_crc_and_game_names_and_system
     connection: &mut SqliteConnection,
     size: u64,
     crc: &str,
-    game_names: Vec<&str>,
+    game_names: &Vec<&str>,
     system_id: i64,
 ) -> Vec<Rom> {
     let size = i64::try_from(size).unwrap();
@@ -1456,6 +1533,46 @@ pub async fn find_roms_without_romfile_by_size_and_crc_and_game_names_and_system
             panic!(
             "Error while finding roms with size {} and CRC {} and game names {:?} and system id {}",
             size, crc, game_names, system_id
+        )
+        })
+}
+
+pub async fn find_roms_without_romfile_by_name_and_size_and_crc_and_game_names_and_system_id(
+    connection: &mut SqliteConnection,
+    name: &str,
+    size: u64,
+    crc: &str,
+    game_names: &Vec<&str>,
+    system_id: i64,
+) -> Vec<Rom> {
+    let size = i64::try_from(size).unwrap();
+    let crc = crc.to_lowercase();
+    let sql = format!(
+        "
+        SELECT r.id, r.name, r.bios, r.size, r.crc, r.md5, r.sha1, r.rom_status, r.game_id, r.romfile_id, r.parent_id
+        FROM roms AS r
+        JOIN games AS g ON r.game_id = g.id
+        WHERE r.romfile_id IS NULL
+        AND r.name = '{}'
+        AND r.size = {}
+        AND r.crc = '{}'
+        AND g.name IN ({})
+        AND g.system_id = {}
+        ORDER BY r.name
+        ",
+        name,
+        size,
+        crc,
+        game_names.iter().map(|game_name| format!("'{}'", game_name)).join(","),
+        system_id,
+    );
+    sqlx::query_as::<_, Rom>(&sql)
+        .fetch_all(connection)
+        .await
+        .unwrap_or_else(|_| {
+            panic!(
+            "Error while finding roms with name {} and size {} and CRC {} and game names {:?} and system id {}",
+            name, size, crc, game_names, system_id
         )
         })
 }